requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.3
cryptography==2.6.1
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
ncclient==0.6.4
netmiko==2.3.3
paramiko==2.4.2
pyang==1.7.8
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyserial==3.4
PyYAML==5.1
scp==0.13.2
six==1.12.0
textfsm==0.4.1
xmltodict==0.12.0
